move necorelease to another file
